db operations

+++ src/models/loans.rs
@@ -1,17 +1,17 @@
 use crate::schema;
-use chrono::serde::{Deserialize, Serialize};
 use diesel::prelude::*;
+use serde::{Deserialize, Serialize};
 
-#[derive(Queryable, Selectable, Serialize, Deserialize)]
+#[derive(Queryable, Selectable, Debug, Serialize, Deserialize)]
 #[diesel(table_name = schema::loans)]
 #[diesel(check_for_backend(diesel::pg::Pg))]
 pub struct Loan {
-    pub id: u32,
+    pub id: i32,
     pub loan_name: String,
-    pub loan_amount: f32,
-    pub interest_rate: f32,
-    pub repayment_period: u32,
-    pub lender_id: u32,
+    pub loan_amount: f64,
+    pub interest_rate: f64,
+    pub repayment_period: i32,
+    pub lender_id: i32,
     pub updated_at: chrono::NaiveDateTime,
     pub created_at: chrono::NaiveDateTime,
 }
@@ -21,8 +21,8 @@ pub struct Loan {
 #[diesel(check_for_backend(diesel::pg::Pg))]
 pub struct NewLoan {
     pub loan_name: String,
-    pub loan_amount: f32,
-    pub interest_rate: f32,
-    pub repayment_period: u32,
-    pub lender_id: u32,
+    pub loan_amount: f64,
+    pub interest_rate: f64,
+    pub repayment_period: i32,
+    pub lender_id: i32,
 }
